backend/requirements.railway.txt
fastapi==0.115.0
uvicorn==0.34.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
openai==1.57.0
numpy==1.26.4
pydantic==2.10.0
python-multipart==0.0.19
requests==2.32.3
